UNST-9983 Use fixed width primitive types to avoid size mismatches in language interop

+++ src/utils_lgpl/dflowfm_io/dflowfm_io_api/wrappers/python/dflowfm_io.py
@@ -71,8 +71,8 @@ class Severity(IntEnum):
 
 class _MduIssue(ctypes.Structure):
     _fields_ = [
-        ("line_number", ctypes.c_int),
-        ("severity", ctypes.c_int),
+        ("line_number", ctypes.c_int32),
+        ("severity", ctypes.c_int32),
         ("message", ctypes.c_char_p),
     ]
 
@@ -116,7 +116,7 @@ class MduDocument:
 
     def load_from_lines(self, data: list[str]) -> None:
         encoded = "\n".join(data).encode("utf-8")
-        _check_result(_lib.mdu_document_load_from_string(self._ref.handle, encoded, len(encoded)))
+        _check_result(_lib.mdu_document_load_from_string(self._ref.handle, encoded, ctypes.c_uint64(len(encoded))))
 
     def save_to_file(self, filename: str) -> None:
         _check_result(_lib.mdu_document_save_to_file(self._ref.handle, filename.encode("utf-8")))
@@ -133,7 +133,7 @@ class MduReport:
 
     def get_issues(self) -> list[Issue]:
         array_out = ctypes.POINTER(_MduIssue)()
-        size_out = ctypes.c_size_t()
+        size_out = ctypes.c_uint64()
         _check_result(_lib.mdu_report_get_issue_list(self._ref.handle, ctypes.byref(array_out), ctypes.byref(size_out)))
         issues = []
         for i in range(size_out.value):
@@ -156,17 +156,17 @@ class MduModel:
         self._ref = ref
 
     def get_dummy_value(self) -> int:
-        value = ctypes.c_int()
+        value = ctypes.c_int32()
         _check_result(_lib.mdu_model_get_dummy_value(self._ref.handle, ctypes.byref(value)))
         return value.value
 
     def get_int(self, key: str) -> int:
-        value = ctypes.c_int()
+        value = ctypes.c_int32()
         _check_result(_lib.mdu_model_get_int(self._ref.handle, key.encode("utf-8"), ctypes.byref(value)))
         return value.value
 
     def get_bool(self, key: str) -> bool:
-        value = ctypes.c_int()
+        value = ctypes.c_int32()
         _check_result(_lib.mdu_model_get_bool(self._ref.handle, key.encode("utf-8"), ctypes.byref(value)))
         return value.value != 0
 
@@ -191,33 +191,33 @@ class MduModel:
         return datetime.fromtimestamp(epoch_out.value, tz=timezone.utc)
 
     def get_enum(self, key: str) -> int:
-        value = ctypes.c_int()
+        value = ctypes.c_int32()
         _check_result(_lib.mdu_model_get_enum(self._ref.handle, key.encode("utf-8"), ctypes.byref(value)))
         return value.value
 
     def get_string_list(self, key: str) -> list[str]:
         array_out = ctypes.POINTER(ctypes.c_char_p)()
-        size_out = ctypes.c_size_t()
+        size_out = ctypes.c_uint64()
         _check_result(_lib.mdu_model_get_string_list(self._ref.handle, key.encode("utf-8"), ctypes.byref(array_out), ctypes.byref(size_out)))
         return [array_out[i].decode("utf-8") for i in range(size_out.value)]
 
     def get_path_list(self, key: str) -> list[Path]:
         array_out = ctypes.POINTER(ctypes.c_char_p)()
-        size_out = ctypes.c_size_t()
+        size_out = ctypes.c_uint64()
         _check_result(_lib.mdu_model_get_path_list(self._ref.handle, key.encode("utf-8"), ctypes.byref(array_out), ctypes.byref(size_out)))
         return [Path(array_out[i].decode("utf-8")) for i in range(size_out.value)]
 
     def get_double_list(self, key: str) -> list[float]:
         array_out = ctypes.POINTER(ctypes.c_double)()
-        size_out = ctypes.c_size_t()
+        size_out = ctypes.c_uint64()
         _check_result(_lib.mdu_model_get_double_list(self._ref.handle, key.encode("utf-8"), ctypes.byref(array_out), ctypes.byref(size_out)))
         return [array_out[i] for i in range(size_out.value)]
 
     def set_int(self, key: str, value: int) -> None:
-        _check_result(_lib.mdu_model_set_int(self._ref.handle, key.encode("utf-8"), ctypes.c_int(value)))
+        _check_result(_lib.mdu_model_set_int(self._ref.handle, key.encode("utf-8"), ctypes.c_int32(value)))
 
     def set_bool(self, key: str, value: bool) -> None:
-        _check_result(_lib.mdu_model_set_bool(self._ref.handle, key.encode("utf-8"), ctypes.c_int(1 if value else 0)))
+        _check_result(_lib.mdu_model_set_bool(self._ref.handle, key.encode("utf-8"), ctypes.c_int32(1 if value else 0)))
 
     def set_double(self, key: str, value: float) -> None:
         _check_result(_lib.mdu_model_set_double(self._ref.handle, key.encode("utf-8"), ctypes.c_double(value)))
@@ -233,18 +233,18 @@ class MduModel:
         _check_result(_lib.mdu_model_set_datetime(self._ref.handle, key.encode("utf-8"), ctypes.c_int64(epoch)))
 
     def set_enum(self, key: str, value: int) -> None:
-        _check_result(_lib.mdu_model_set_enum(self._ref.handle, key.encode("utf-8"), ctypes.c_int(value)))
+        _check_result(_lib.mdu_model_set_enum(self._ref.handle, key.encode("utf-8"), ctypes.c_int32(value)))
 
     def set_string_list(self, key: str, values: list[str]) -> None:
         encoded = [v.encode("utf-8") for v in values]
         arr = (ctypes.c_char_p * len(encoded))(*encoded)
-        _check_result(_lib.mdu_model_set_string_list(self._ref.handle, key.encode("utf-8"), arr, ctypes.c_size_t(len(encoded))))
+        _check_result(_lib.mdu_model_set_string_list(self._ref.handle, key.encode("utf-8"), arr, ctypes.c_uint64(len(encoded))))
 
     def set_path_list(self, key: str, values: list[Path | str]) -> None:
         encoded = [str(v).encode("utf-8") for v in values]
         arr = (ctypes.c_char_p * len(encoded))(*encoded)
-        _check_result(_lib.mdu_model_set_path_list(self._ref.handle, key.encode("utf-8"), arr, ctypes.c_size_t(len(encoded))))
+        _check_result(_lib.mdu_model_set_path_list(self._ref.handle, key.encode("utf-8"), arr, ctypes.c_uint64(len(encoded))))
 
     def set_double_list(self, key: str, values: list[float]) -> None:
         arr = (ctypes.c_double * len(values))(*values)
-        _check_result(_lib.mdu_model_set_double_list(self._ref.handle, key.encode("utf-8"), arr, ctypes.c_size_t(len(values))))
+        _check_result(_lib.mdu_model_set_double_list(self._ref.handle, key.encode("utf-8"), arr, ctypes.c_uint64(len(values))))
